Better url bulder, transport test harness
